match expected clock and reset names (+polarity)
Chisel circuits always feature a "clock"
and a "reset" input. This assumption is baked
into rfuzz. Thus we need to change the name
on the toplevel aes_core module.

Also important: Chisel expects a high-active
reset whereas the aes_core is written with a
low-active reset in mind.
We thus need to invert the input to the
toplevel module.